Refactor folder structure: add folders for BookmarksPage and CrimeDataPage
